+++ src/main/groovy/org/codenarc/rule/formatting/ClassStartsWithBlankLineRule.groovy
@@ -24,14 +24,22 @@ import org.codenarc.util.AstUtil
 
 /**
  * Check whether the class starts with a blank line. By default, it enforces that there must be a blank line after
- * the opening class brace, except if the class is empty and is written in a single line. A blank line is defined as
- * any line that does not contain any visible characters. This rule can be configured with the following properties:
+ * the opening class brace, except:
+ *   <ul>
+ *     <li>If the class is synthetic (generated)</li>
+ *     <li>If the class is empty and is written in a single line</li>
+ *     <li>If the class is a Script class</li>
+ *   <ul>
+ * <p>
+ * A blank line is defined as any line that does not contain any visible characters.
+ * <p>
+ * This rule can be configured with the following properties:
  * <ul>
- *  <li><i>ignoreSingleLineClasses</i>: a boolean property to forbid single line classes.
- *  If it is false, then single line classes are considered a violation. Default value is true</li>
- *  <li><i>blankLineRequired</i>: a boolean property to define if there may be a blank line after the opening class
- *  brace. If it is false, the first content after the brace must not be a blank line. Otherwise, it must be a blank
- *  line. Default value is true</li>
+ *   <li><i>ignoreSingleLineClasses</i>: a boolean property to forbid single line classes.
+ *     If it is false, then single line classes are considered a violation. Default value is true</li>
+ *   <li><i>blankLineRequired</i>: a boolean property to define if there may be a blank line after the opening class
+ *     brace. If it is false, the first content after the brace must not be a blank line. Otherwise, it must be a blank
+ *     line. Default value is true</li>
  * <ul>
  *
  * @author David Ausín
@@ -52,6 +60,8 @@ class ClassStartsWithBlankLineAstVisitor extends AbstractAstVisitor {
     @Override
     protected void visitClassComplete(ClassNode classNode) {
         if (classNode.lineNumber == -1) { return }
+
+        if (classNode.isScript()) { return }
 
         if (isSingleLineClassViolation() && isSingleLineClass(classNode)) { return }