+++ src/test/groovy/org/codenarc/rule/formatting/ClassStartsWithBlankLineRuleTest.groovy
@@ -566,6 +566,23 @@ class ClassStartsWithBlankLineRuleTest extends AbstractRuleTestCase<ClassStartsW
         assertNoViolations(SOURCE)
     }
 
+    @Test
+    void testNoViolationsForScriptClass() {
+        final String SOURCE = '''
+            job('test-job') {
+                triggers { // fails here, no new line above
+                    cron('* * * * *')
+                }
+            
+                steps {
+                    groovyScriptFile('my_script.groovy')
+                }
+            }
+        '''
+
+        assertNoViolations(SOURCE)
+    }
+
     @Override
     protected ClassStartsWithBlankLineRule createRule() {
         new ClassStartsWithBlankLineRule()